Make the code cleaner and easier to maintain
